updates to templates thus far are now in templates dir
