--- openai-asst-streaming-js/package.json
@@ -1,14 +1,0 @@
-{
-    "name": "openai-asst-streaming",
-    "version": "1.0.0",
-    "description": "",
-    "main": "Main.js",
-    "scripts": {
-        "test": "echo \"Error: no test specified\" && exit 1"
-    },
-    "author": "",
-    "license": "MIT",
-    "dependencies": {
-        "openai": "^4.0.0"
-    }
-}
