+++ src/ai/.x/templates/openai-chat-w-asst-w-streaming-js/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "openai-asst-streaming",
+  "version": "1.0.0",
+  "description": "",
+  "main": "Main.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "MIT",
+  "dependencies": {
+    "openai": "^4.0.0"
+  }
+}
