+++ src/ai/.x/templates/openai-chat-w-asst-webpage-js/.env
@@ -1,0 +1,10 @@
+<#@ template hostspecific="true" #>
+<#@ output extension=".env" encoding="utf-8" #>
+<#@ parameter type="System.String" name="AZURE_OPENAI_ENDPOINT" #>
+<#@ parameter type="System.String" name="AZURE_OPENAI_KEY" #>
+<#@ parameter type="System.String" name="AZURE_OPENAI_CHAT_DEPLOYMENT" #>
+<#@ parameter type="System.String" name="AZURE_OPENAI_SYSTEM_PROMPT" #>
+AZURE_OPENAI_CHAT_DEPLOYMENT=<#= AZURE_OPENAI_CHAT_DEPLOYMENT #>
+AZURE_OPENAI_KEY=<#= AZURE_OPENAI_KEY #>
+AZURE_OPENAI_ENDPOINT=<#= AZURE_OPENAI_ENDPOINT #>
+AZURE_OPENAI_SYSTEM_PROMPT=<#= AZURE_OPENAI_SYSTEM_PROMPT #>
