+++ src/ai/.x/templates/openai-chat-w-asst-webpage-js/_.json
@@ -1,0 +1,14 @@
+{
+  "_LongName": "OpenAI Chat w/ Assistants Webpage",
+  "_ShortName": "openai-chat-w-asst-webpage",
+  "_Language": "JavaScript",
+  "ClassName": "OpenAIAssistantsStreamingClass",
+  "AZURE_OPENAI_SYSTEM_PROMPT": "You are a helpful AI assistant.",
+  "AZURE_OPENAI_API_KEY": "<insert your Azure OpenAI API key here>",
+  "AZURE_OPENAI_API_VERSION": "<insert your Azure OpenAI API version here>",
+  "AZURE_OPENAI_ENDPOINT": "<insert your Azure OpenAI endpoint here>",
+  "AZURE_OPENAI_CHAT_DEPLOYMENT": "<insert your Azure OpenAI chat deployment name here>",
+  "OPENAI_API_KEY": "<insert your OpenAI API key here>",
+  "OPENAI_ORG_ID": "<insert your OpenAI organization ID here>",
+  "OPENAI_MODEL_NAME": "<insert your OpenAI model name here>"
+}
