+++ src/ai/.x/templates/openai-chat-w-asst-webpage-js/index.html
