+++ src/ai/.x/templates/openai-chat-w-asst-webpage-js/package.json
