+++ src/ai/.x/templates/openai-chat-w-asst-webpage-js/webpack.config.js
